Fix MSVC /W4 warnings in common library and kernels
Resolve all MSVC warnings that fire under /W4 /WX in the core library
(common/) and BVH kernels (kernels/). Changes are purely mechanical
renames and casts — no logic is altered.

Warning categories fixed:
  C4189  unused initialized local variable (alloc.h: 'done' only live
         inside assert, which is a no-op in Release; add (void)done)
  C4245  signed/unsigned mismatch (ray.mask = -1 -> 0xFFFFFFFFu, and
         various index/count assignments)
  C4456  local declaration hides outer local (loop-variable shadowing)
  C4457  local declaration hides function parameter
  C4458  local declaration hides class member (lambda captures, etc.)
  C4459  local declaration hides global (parameter named same as global)

+++ kernels/bvh/bvh_node_ref.h
@@ -66,7 +66,7 @@ namespace embree
     static const size_t byteNodeAlignment = 4*N;
 
     /*! highest address bit is used as barrier for some algorithms */
-    static const size_t barrier_mask = (1LL << (8*sizeof(size_t)-1));
+    static const size_t barrier_mask = (size_t(1) << (8*sizeof(size_t)-1));
 
     /*! Masks the bits that store the number of items per leaf. */
     static const size_t align_mask = byteAlignment-1;
